refactor(engine):  extract evaluation into distinct class
